Improves some of the example code and fixes typo
In order to make the documentation more clear, this commit changes the
example of how the dynamic_print method is used. This will make it more
clear about what is going on in the examples instead of being formatted
like a test.
This commit also fixes a typo for the LengthErrorData struct
documentation.