Adds methods for comparing internal grid sizes
In order to make the printer easier to work with, this commit adds two
methods for comparing internally stored grids. One returns the size of
the grid, the other determines if a passed in grid will be valid or not.

In addition, this commit cleans a bit of the code and removes guard.
Guard is now a method implemented into the standard library. On top of
that, guard does not build in nightly mode due to some useless legacy
code that they've left in the project forever now.