Adds options for where to print grids with the dynamic_print method
In order to have an easier time using the printer, this commit adds
suppose for preset locations for grids to be print to the screen.
There are 9 total options from a set of [Top, Middle, Bottom] on the
Y-axis, and [Left, Middle, Right] for the X-axis.
Currently The only functioning option for the X-axis is Left. This is
due to the fact that newlines are accounted for when building out pixel
differences. These will need to be replaced with escape codes upon
building the printable difference.