Merges printable diff and pixel diff methods, fixes tests, and more
In order to remove unnecessary steps, this commit skips the steps of
storing differences of pixels, and directly builds the escape codes with
the printable difference.

In addition, this commit:
Fixes all tests that broke as a result of any changes.
Adds additional quality of life methods.
Changes the name of PrintingOptions to PrintingPositions.
Fully implements PrintingPositions.
Adds an extra example using PrintingPositions.