Fixes bugs when clearing and changing the size of the grid to be printed
In order to fix some bugs, this commit moves any instance where the
position of the grid to be printed may change, to it's their methods.
That would mean anytime the internal dimensions or origin for the grid.
If either of those would change, the
`printing_position_changed_since_last_print` flag is set to true.

In addition, this commit fixes a bug where the `clear_grid` method
swapped the width and height of the grid when creating the grid of
whitespace to be printed over it.
Along with that, this commit fixes another bug with the
`replace_currently_printed_grid` method. Where the origin was not being
updated if it already existed.