Adds a methods to make non-rectangular grids rectangular and other.
In order to make the creations of grids for this crate easier, this
commit adds a method that will add whitespace to every row of a string
so that they match the largest row.

In addition, this commit makes the method for getting a grid's
dimensions public, and adds an additional method for checking if a grid
is a valid rectangle.